odm-2-0/requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
daff==1.4.2
dbt-adapters==1.16.3
dbt-common==1.27.1
dbt-core==1.10.4
dbt-duckdb==1.9.4
dbt-extractor==0.6.0
dbt-protos==1.0.348
dbt-semantic-interfaces==0.9.0
deepdiff==7.0.1
duckdb==1.3.2
idna==3.10
importlib_metadata==8.7.0
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
leather==0.4.0
lxml==6.0.0
MarkupSafe==3.0.2
mashumaro==3.14
more-itertools==10.7.0
msgpack==1.1.1
networkx==3.4.2
numpy==2.2.6
ordered-set==4.1.0
packaging==25.0
pandas==2.3.1
parsedatetime==2.6
pathspec==0.12.1
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
snowplow-tracker==1.1.0
sqlparse==0.5.3
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
